Fix CodeQL cs/loss-of-precision alerts 715-718: widen int operands before multiply
Widen an int operand to double before multiplying, rather than only
widening the result, so the multiplication itself cannot overflow as
32-bit int arithmetic before being stored in a double-typed field.

- OrcaSlicingPipelineService.cs:1257 - PrintTimeSeconds computation
- JobSchedulingService.cs:811 - Weekly recurrence AddDays computation
- GcodeMetadataExtractorService.cs:632 - TryParsePrintTime minutes computation

Part of #1380 (CodeQL Tier 4 quality-rule campaign, batch 30).

Copilot-Session: 39577b36-cbbd-4ee1-8a97-e85996c81141

+++ src/orcaslicer-worker/Services/OrcaSlicingPipelineService.cs
@@ -1254,7 +1254,9 @@ public partial class OrcaSlicingPipelineService : ISlicingPipelineService
             Match tm = printTimeRegex.Match(line);
             if (tm.Success)
             {
-                metadata.PrintTimeSeconds = (int.Parse(tm.Groups[1].Value, System.Globalization.CultureInfo.InvariantCulture) * 3600) + (int.Parse(tm.Groups[2].Value, System.Globalization.CultureInfo.InvariantCulture) * 60);
+                // Multiply in double space (not int) before assigning to the double PrintTimeSeconds field,
+                // so an unexpectedly large parsed hour/minute value cannot silently overflow as int arithmetic.
+                metadata.PrintTimeSeconds = ((double)int.Parse(tm.Groups[1].Value, System.Globalization.CultureInfo.InvariantCulture) * 3600) + ((double)int.Parse(tm.Groups[2].Value, System.Globalization.CultureInfo.InvariantCulture) * 60);
             }
             else
             {
